Added a delete method and delete controller, changed the HTTP verbs in the retrieval controller
Introduced a method, that deletes a project record in the database, DeleteProject(). It deletes both from the projects table and projectusers table. It needs to delete from both, as otherwise it violates the constrains of FK between the two tables.  Respectively, introduced a deletion controller, that handles the deletion of projects.
Finally changed the HTTP verb in the retrieval controller to be get. Now the controller expects all routes to include parameters in the url.