Registered SupabaseClient and ProjetService as services and prepared them for DI + Test flow
Registered the SupabaseClient as a singleton in the middleware (Program.cs) and prepared it to be dependency injected into constructors where needed. Currently, the SupabaseClient singleton is injected only in the ProjectService.
Registered ProjectService as a scoped service in the middleware (Program.cs) and was also prepared to be dependency injected into constructors where needed. The ProjectService is injected into the controllers.
The controllers were also re-factored and already follow the pattern of "api/v-/[controller-name]/route"
Re-factored the test project and set up the testing flow again.